download.pl: copy from local mirrors only if dl file not present
In recent commits, we have added forced download call and added
check to confirm if the hash is correct so that we dont end up
using an incorrect/corrupt tarball.
This change ensures that we are not copying from local mirror during
the forced download if the file is already present.
This change also ensures that corruption check is done on the files
copied from the local mirror locations.

Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>
Change-Id: Ia6fac34a43d9893f79e1f92b022c90040fe9cd16