busybox: remove nslookup_lede/openwrt.patch
The nslookup_lede/openwrt applet was introduced in de5b8e5. It was
introduced because:

  Add a new LEDE nslookup applet which is compatible with musl libc
  and providing more features like ability to specify query type.

  In contrast to busybox' builtin nslookup applet, this variant does
  not rely on libc resolver internals but uses explicit send logic
  and the libresolv primitives to parse received DNS responses.

In busybox this applet is added in 0dd3be8. In particular, this commit
introduces the variable NSLOOKUP_BIG. We set the default to true and
so nothing changes.

Signed-off-by: Nick Hainke <vincent@systemli.org>
(cherry picked from commit b36b8b6929c6d6b17edddfb4597cf6a26a991ed0)

Change-Id: I34caecd3ce67feca2111b163f167d1f3e2593e2b
Signed-off-by: Ram Chandra Jangir <quic_rjangir@quicinc.com>