ipq95xx: sysupgrade: upgrade both partitions for TZ,TME and SBL
During sysupgrade the following images are to be
updated in both the primary and alternate partitions
on the flash.
1.	SBL
2.	TME
3.	QSEE

These images are always expected to boot from the
primary partition upon sysupgrade. As the above
images will have same versions on both the partitions,
their BOOTCONFIG entries are removed, as it may
differ from the actual primary_boot value shown as
0 or 1.

After sysupgrade, the image versions could be
verified through the version IDs from sysfs entries
as in sys/firmware/devicetree/base/*version

Change-Id: Ie3f0bc3f31fa55a74010b3a4f507780902b1df6b
Signed-off-by: Sandhya KS <sandhyks@codeaurora.org>