scripts : Generate SOURCE_DATE_EPOCH with current date time
Currently SOURCE_DATE_EPOCH is set by default to the date time
of last git commit in Openwrt to make the builds reproducible

Kernel version time in uname and /proc/version is set using
this SOURCE_DATE_EPOCH. Uboot timestamp also depends on the
same mechanism.

To get the actual build date and time in uboot and kernel, we
need to override the default SOURCE_DATE_EPOCH logic by setting
the current date time to a file "version.date" in the buildroot
root dir before running make as required by the script. This
overriding will be gaurded by an environment variable -
QSDK_OVERRIDE_GIT_TIMESTAMP. If this is set in the build
environment, then the git timestamp will be overridden and
the build will take the current date time, otherwise, the
image will still use the commit date time of the last openwrt
commit.

If QSDK_OVERRIDE_GIT_TIMESTAMP is set to a non-zero value:
	Build timestamp will be current date-time
Else:
	Build timestamp will be derived from the output of
	'git log -1 --format=format:%ct' in OpenWRT tree

Change-Id: I01cc8372b232712c7dbfef81dcfc3008c816cca0
Signed-off-by: Sandhya KS <sandhyks@codeaurora.org>