Zlib: Fix for CVE-2016-9840
This change is based on the upstream commmit,
https://github.com/madler/zlib/commit/6a043145ca6e9c55184013841a67b2fef87e44c0

Remove offset pointer optimization in inftrees.c.

inftrees.c was subtracting an offset from a pointer to an array,
in order to provide a pointer that allowed indexing starting at
the offset. This is not compliant with the C standard, for which
the behavior of a pointer decremented before its allocated memory
is undefined. Per the recommendation of a security audit of the
zlib code by Trail of Bits and TrustInSoft, in support of the
Mozilla Foundation, this tiny optimization was removed, in order
to avoid the possibility of undefined behavior.

Change-Id: Iaa89a083b2b772173eeb839c23c1826f1c675ff5
Signed-off-by: Sanjeev N <quic_nsanjeev@quicinc.com>